requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-crispy-forms==1.13.0
django-formtools==2.3
django-otp==1.1.3
django-otp-yubikey==1.0.1
django-phonenumber-field==6.0.0
django-two-factor-auth==1.13.2
Django-Verify-Email==1.0.9
idna==3.3
phonenumbers==8.12.41
pycryptodome==3.13.0
PyJWT==2.3.0
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
sqlparse==0.4.2
twilio==7.5.0
urllib3==1.26.8
YubiOTP==1.0.0
